+++ tests/backend/patches/lora_conversions/test_flux_kohya_lora_conversion_utils.py
@@ -4,22 +4,25 @@ import torch
 
 from invokeai.backend.flux.model import Flux
 from invokeai.backend.flux.util import params
-from invokeai.backend.patches.conversions.flux_kohya_lora_conversion_utils import (
+from invokeai.backend.patches.lora_conversions.flux_kohya_lora_conversion_utils import (
     _convert_flux_transformer_kohya_state_dict_to_invoke_format,
     is_state_dict_likely_in_flux_kohya_format,
     lora_model_from_flux_kohya_state_dict,
 )
-from invokeai.backend.patches.conversions.flux_lora_constants import FLUX_LORA_CLIP_PREFIX, FLUX_LORA_TRANSFORMER_PREFIX
-from tests.backend.patches.conversions.lora_state_dicts.flux_lora_diffusers_format import (
+from invokeai.backend.patches.lora_conversions.flux_lora_constants import (
+    FLUX_LORA_CLIP_PREFIX,
+    FLUX_LORA_TRANSFORMER_PREFIX,
+)
+from tests.backend.patches.lora_conversions.lora_state_dicts.flux_lora_diffusers_format import (
     state_dict_keys as flux_diffusers_state_dict_keys,
 )
-from tests.backend.patches.conversions.lora_state_dicts.flux_lora_kohya_format import (
+from tests.backend.patches.lora_conversions.lora_state_dicts.flux_lora_kohya_format import (
     state_dict_keys as flux_kohya_state_dict_keys,
 )
-from tests.backend.patches.conversions.lora_state_dicts.flux_lora_kohya_with_te1_format import (
+from tests.backend.patches.lora_conversions.lora_state_dicts.flux_lora_kohya_with_te1_format import (
     state_dict_keys as flux_kohya_te1_state_dict_keys,
 )
-from tests.backend.patches.conversions.lora_state_dicts.utils import keys_to_mock_state_dict
+from tests.backend.patches.lora_conversions.lora_state_dicts.utils import keys_to_mock_state_dict
 
 
 @pytest.mark.parametrize("sd_keys", [flux_kohya_state_dict_keys, flux_kohya_te1_state_dict_keys])
